Move group related files under new directory
