+++ timApp/static/scripts/tim/plugin/group-dashboard/group-name.component.scss
